update, delete, comment

+++ myMovieReviews/movie_record/movies/views.py
@@ -10,7 +10,8 @@ def movie_create(request):
     if request.method == 'POST':
         form = MovieForm(request.POST)
         if form.is_valid():
-            form.save()
+            movie = form.save(commit=False)
+            movie.save()
             return redirect('movie_list')
     else:
         form = MovieForm()
@@ -35,8 +36,10 @@ def movie_update(request, pk):
 
 def movie_delete(request, pk):
     movie = get_object_or_404(Movie, pk=pk)
-    movie.delete()
-    return redirect('movie_list')
+    if request.method == 'POST':
+        movie.delete()
+        return redirect('movie_list')
+    return render(request, 'movies/movie_confirm_delete.html', {'movie': movie})
 
 def add_comment(request, pk):
     movie = get_object_or_404(Movie, pk=pk)
